refactoring & standardization: Renamed module _io to _dio since it deals with Data Input Output
