+++ src/astrohack/_utils/_dio.py
